requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
asyncio==3.4.3
attrs==24.2.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
discord.py==2.4.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.55.8
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.25.2
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2024.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.20.1
torch==2.4.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.17.2
yarl==1.13.1
